Move the config objects to their own directory where we can control their rules better
